move the color handling code into a separate module
