+++ blackdoc/__main__.py
@@ -1,9 +1,7 @@
 import argparse
 import datetime
 import difflib
-import functools
 import pathlib
-import re
 import sys
 
 import black
@@ -15,10 +13,8 @@ from .blackcompat import (
     get_gitignore,
     normalize_path_maybe_ignore,
     read_pyproject_toml,
-    wrap_stream_for_windows,
 )
-
-colors_re = re.compile("\033" + r"\[[0-9]+(?:;[0-9]+)*m")
+from .colors import color_diff, colorize, err, out
 
 
 def check_format_names(string):
@@ -75,60 +71,6 @@ def collect_files(src, include, exclude, extend_exclude, force_exclude, quiet, v
             yield path
         else:
             print(f"invalid path: {path}", file=sys.stderr)
-
-
-def colorize(string, fg=None, bold=False):
-    foreground_colors = {
-        "white": 37,
-        "cyan": 36,
-        "green": 32,
-        "red": 31,
-    }
-    bold_code = 1
-    reset_code = 0
-
-    codes = []
-    if bold:
-        codes.append(bold_code)
-
-    if fg:
-        codes.append(foreground_colors.get(fg, fg))
-
-    return f"\033[{';'.join(map(str, codes))}m{string}\033[{reset_code}m"
-
-
-def remove_colors(message):
-    return "".join(colors_re.split(message))
-
-
-# signature inspired by click.secho
-def custom_print(message, end="\n", file=sys.stdout, **styles):
-    if file.isatty():
-        message = colorize(message, **styles)
-    else:
-        message = remove_colors(message)
-
-    print(message, end=end, file=wrap_stream_for_windows(file))
-
-
-out = functools.partial(custom_print, file=sys.stdout)
-err = functools.partial(custom_print, file=sys.stderr)
-
-
-def color_diff(contents):
-    """Inject the ANSI color codes to the diff."""
-    lines = contents.split("\n")
-    for i, line in enumerate(lines):
-        if line.startswith("+++") or line.startswith("---"):
-            line = colorize(line, fg="white", bold=True)  # bold white, reset
-        elif line.startswith("@@"):
-            line = colorize(line, fg="cyan")  # cyan, reset
-        elif line.startswith("+"):
-            line = colorize(line, fg="green")  # green, reset
-        elif line.startswith("-"):
-            line = colorize(line, fg="red")  # red, reset
-        lines[i] = line
-    return "\n".join(lines)
 
 
 def unified_diff(a, b, path, color):
